+++ src/components/Post/LikeButton/index.js
@@ -1,8 +1,10 @@
 import { LikeButtonWrapper } from "./style";
 
-export default function LikeButton({ toggleSelection, likedByMe}) {
+export default function LikeButton({toggleSelection, likedByMe,	setLikedByMe}) {
+	
 
 	return (
+	
 		<LikeButtonWrapper selected={likedByMe} onClick={toggleSelection}>
 			{likedByMe ? <ion-icon name="heart"></ion-icon> : <ion-icon name="heart-outline"></ion-icon>}
 		</LikeButtonWrapper>
